Answers

+++ app/controllers/questions_controller.rb
@@ -8,6 +8,7 @@ class QuestionsController < ApplicationController
   
   def show
   	@question = Question.find(params[:id])
+    @answer = Answer.new
   end
 
   def new
@@ -16,6 +17,7 @@ class QuestionsController < ApplicationController
 
   def create
   	@question = Question.new(question_params)
+    @question[:user_id] = current_user[:id]
 		if @question.save
 			flash[:success] = "Pregunta agregada exitosamente!"
 			redirect_to questions_path  	
@@ -27,32 +29,47 @@ class QuestionsController < ApplicationController
 
   def edit
   	@question = Question.find(params[:id])
+    if @question[:user_id] == current_user[:id] 
+    else
+      flash[:danger] = "Usted no esta autorizado para editar esta pregunta"
+      redirect_to question_path(@question[:id])
+    end
   end
 
   def update
   	@question = Question.find(params[:id])
-  	if @question.update(question_params)
-  		flash[:success] = "Pregunta modificada exitosamente!"
-  		redirect_to questions_path
-  	else
-  		flash[:danger] = "Ha ocurrido un error"
-  		render :edit
-  	end
+    if @question[:user_id] == current_user[:id] 
+    	if @question.update(question_params)
+    		flash[:success] = "Pregunta modificada exitosamente!"
+    		redirect_to questions_path
+    	else
+    		flash[:danger] = "Ha ocurrido un error"
+    		render :edit
+    	end
+    else
+      flash[:danger] = "Usted no esta autorizado para editar esta pregunta"
+      redirect_to question_path(@question[:id])
+    end
   end
 
   def destroy
   	@question = Question.find(params[:id])
-  	if @question.destroy
-  		flash[:success] = "Pregunta eliminada exitosamente!"
-  		redirect_to questions_path
-  	else
-  		flash[:danger] = "Ha ocurrido un error"
-  		render :show
-  	end
+    if @question[:user_id] == current_user[:id] 
+    	if @question.destroy
+    		flash[:success] = "Pregunta eliminada exitosamente!"
+    		redirect_to questions_path
+    	else
+    		flash[:danger] = "Ha ocurrido un error"
+    		render :show
+    	end
+    else
+      flash[:danger] = "Usted no esta autorizado para eliminar esta pregunta"
+      redirect_to question_path(@question[:id])
+    end
   end
 
   private
     def question_params
-      params.require(:question).permit(:title, :description)
+      params.require(:question).permit(:title, :description, :user_id)
     end
 end
